Remove rivet game and hi-score/leaderboard feature; trim implementation
Remove the rivet game entirely: crate, embedded assets, web build
output, the index.html card + i18n strings, the service-worker
precache entries (cache bumped to v7), and the workspace member.

Remove the Supabase-backed hi-score/leaderboard feature: the initials
overlay, leaderboard JS bridge, SQL schema, and Playwright E2E suite.
draw_hud drops the HI field; Session no longer tracks hi-score.

Trim the implementation for readability: collapse a manual clamp,
use is_multiple_of, drop redundant casts, and replace a hand-rolled
find loop with an iterator. Build is warning-free on native; web wasm
artifacts rebuilt for all five remaining games.

+++ crates/blip_assets/src/galactic_defender.rs
@@ -94,7 +94,7 @@ fn alien(kind: usize) -> Vec<u8> {
     for row in 0..5 {
         for col in 0..5 {
             if patterns[kind][row] & (1 << (4 - col)) != 0 {
-                let px_x = ox + col as i32 * 2;
+                let px_x = ox + col * 2;
                 let px_y = oy + row as i32 * 2;
                 img.set(px_x,     px_y,     r, g, b);
                 img.set(px_x + 1, px_y,     r, g, b);
